refactor(claude): extract calculateSessionTokens into claude.session-tokens.lib.mjs
claude.lib.mjs was 1401 lines, past the 1350-line early-warning threshold of
the CI file-headroom check (issue #1593).

Session token accounting reads the session JSONL and prices it; it shares no
state with the executor, so it moves to its own module using plain node
built-ins instead of the use-m bootstrap. claude.lib.mjs re-exports it, so
importers are unaffected. The now-unused Decimal and budget-stats imports move
with it.

claude.lib.mjs: 1401 -> 1247 lines.

Refs #2175
